Move `ServerEventChannel` and `ClientEventChannel` into their modules
Initially it was a single type, but then it got splitted, we just forgot
to move it.
