+++ src/core/network_channels.rs
@@ -72,7 +72,8 @@ impl NetworkChannels {
 
     /// Sets the maximum usage bytes for a specific server channel.
     ///
-    /// [`ReplicationChannel`] or [`ServerEventChannel<T>`](crate::network_event::ServerEventChannel) can be passed as `id`.
+    /// [`ReplicationChannel`] or [`ServerEventChannel<T>`](crate::network_event::server_event::ServerEventChannel)
+    /// can be passed as `id`.
     /// Without calling this function, the default value will be used.
     /// See also [`Self::set_default_max_bytes`].
     pub fn set_server_max_bytes(&mut self, id: impl Into<u8>, max_bytes: usize) {
