+++ src/network_event.rs
@@ -1,70 +1,10 @@
 pub mod client_event;
 pub mod server_event;
 
-use std::{marker::PhantomData, time::Duration};
+use std::time::Duration;
 
 use bevy::{ecs::entity::EntityHashMap, prelude::*};
 use bevy_renet::renet::SendType;
-
-/// Holds a server's channel ID for `T`.
-#[derive(Resource)]
-pub struct ServerEventChannel<T> {
-    id: u8,
-    marker: PhantomData<T>,
-}
-
-impl<T> ServerEventChannel<T> {
-    fn new(id: u8) -> Self {
-        Self {
-            id,
-            marker: PhantomData,
-        }
-    }
-}
-
-impl<T> Clone for ServerEventChannel<T> {
-    fn clone(&self) -> Self {
-        *self
-    }
-}
-
-impl<T> Copy for ServerEventChannel<T> {}
-
-impl<T> From<ServerEventChannel<T>> for u8 {
-    fn from(value: ServerEventChannel<T>) -> Self {
-        value.id
-    }
-}
-
-/// Holds a client's channel ID for `T`.
-#[derive(Resource)]
-pub struct ClientEventChannel<T> {
-    id: u8,
-    marker: PhantomData<T>,
-}
-
-impl<T> ClientEventChannel<T> {
-    fn new(id: u8) -> Self {
-        Self {
-            id,
-            marker: PhantomData,
-        }
-    }
-}
-
-impl<T> Clone for ClientEventChannel<T> {
-    fn clone(&self) -> Self {
-        *self
-    }
-}
-
-impl<T> Copy for ClientEventChannel<T> {}
-
-impl<T> From<ClientEventChannel<T>> for u8 {
-    fn from(value: ClientEventChannel<T>) -> Self {
-        value.id
-    }
-}
 
 /// Event delivery guarantee.
 ///
